[Edit] kakao 자동 로그인가능 하도록  로직 변경, VM class 형태로 변경 전
