+++ src/pages/category/CategoryViewModel.ts
