+++ src/pages/menual/Menual.tsx
@@ -1,6 +1,7 @@
 import styled from "styled-components";
+import React from "react";
 
-const Menual = () => {
+const Menual: React.FC = () => {
     return <Header>테스트 설명 페이지</Header>;
 };
